feat: create union type from tailwind theme text color tokens
