+++ react-client/tailwind.theme.ts
@@ -1,0 +1,176 @@
+import type { Config } from "tailwindcss";
+
+const SPACING_DEFAULTS = {
+  0: "0",
+  auto: "auto",
+  xxxxxlarge: "var(--xxxxxlarge)",
+  xxxxlarge: "var(--xxxxlarge)",
+  xxxlarge: "var(--xxxlarge)",
+  xxlarge: "var(--xxlarge)",
+  xlarge: "var(--xlarge)",
+  large: "var(--large)",
+  base: "var(--base)",
+  medium: "var(--medium)",
+  small: "var(--small)",
+  xsmall: "var(--xsmall)",
+};
+
+export const THEME_COLORS = {
+  "brand-main": "var(--purple-300)",
+  "brand-secondary": "var(--orange-300-contrast)",
+  "brand-blue": "var(--brand-blue)",
+  "border-strong": "var(--border-color-strong)",
+  "purple-500": "var(--purple-500)",
+  "purple-300": "var(--purple-300)",
+  "purple-100": "var(--purple-100)",
+  "orange-500": "var(--orange-500)",
+  "orange-300-contrast": "var(--orange-300-contrast)",
+  "orange-300-main": "var(--orange-300-main)",
+  "orange-200": "var(--orange-200)",
+  "orange-100": "var(--orange-100)",
+  "blue-500": "var(--blue-500)",
+  "blue-300": "var(--blue-300)",
+  "blue-100": "var(--blue-100)",
+  "maroon-500": "var(--maroon-500)",
+  "maroon-300": "var(--maroon-300)",
+  "maroon-100": "var(--maroon-100)",
+  "green-500": "var(--green-500)",
+  "success-contrast": "var(--green-500)",
+  "success-background": "var(--green-100)",
+  "green-300": "var(--green-300)",
+  "green-100": "var(--green-100)",
+  "yellow-500": "var(--yellow-500)",
+  "yellow-300": "var(--yellow-300)",
+  "yellow-100": "var(--yellow-100)",
+  "red-500": "var(--red-500)",
+  "red-300": "var(--red-300)",
+  "red-100": "var(--red-100)",
+  "error-contrast": "var(--red-500)",
+  "error-background": "var(--red-100)",
+  "neutral-950": "var(--neutral-950)",
+  "neutral-900": "var(--neutral-900)",
+  "neutral-800": "var(--neutral-800)",
+  "neutral-700": "var(--neutral-700)",
+  "neutral-600": "var(--neutral-600)",
+  "neutral-500": "var(--neutral-500)",
+  "neutral-400": "var(--neutral-400)",
+  "neutral-300": "var(--neutral-300)",
+  "neutral-200": "var(--neutral-200)",
+  "neutral-100": "var(--neutral-100)",
+  "neutral-50": "var(--neutral-50)",
+  "neutral-white": "var(--neutral-white)",
+  "background-disabled": "var(--background-disabled)",
+  "primary-900": "var(--text-primary-900)",
+  "secondary-700": "var(--text-secondary-700)",
+  "tertiary-600": "var(--text-tertiary-600)",
+  main: "var(--text-main)",
+  "border-base-color": "var(--border-base-color)",
+  "border-secondary": "var(--border-secondary)",
+  "text-danger-default": "var(--text-danger-default)",
+  "text-disabled": "var(--text-color-disabled)",
+  "text-hinted": "var(--text-hinted)",
+  "text-color": "var(--text-color)",
+  "icon-danger": "var(--icon-danger)",
+  "text-color-supporting": "var(--text-color-supporting)",
+  "background-color-tertiary": "var(--background-color-tertiary)",
+  white: "#ffffff",
+  black: "#000000",
+  transparent: "transparent",
+};
+
+export const CUSTOM_THEME: Config["theme"] = {
+  boxShadow: {
+    "elevation-1": "0px 2px 4px 0px rgba(0, 0, 0, 0.24)",
+    "elevation-3": "0px 10px 14px 0px rgba(0, 0, 0, 0.08)",
+  },
+  colors: THEME_COLORS,
+  fontFamily: {
+    franklin: "var(--body-font-family)",
+    raleway: "var(--heading-font-family)",
+    roboto: "var(--fixed-font-family)",
+  },
+  fontSize: {
+    xs: "0.75rem",
+    sm: "0.875rem",
+    base: "1rem",
+    xl: "1.25rem",
+    "2xl": "1.5rem",
+    "3xl": "2rem",
+  },
+  gap: SPACING_DEFAULTS,
+  margin: SPACING_DEFAULTS,
+  padding: SPACING_DEFAULTS,
+  space: SPACING_DEFAULTS,
+  translate: SPACING_DEFAULTS,
+  extend: {
+    animation: {
+      fadeIn: "fadeIn 300ms ease-in-out",
+      fadeOut: "fadeOut 300ms ease-in-out",
+      slideInFromBottom: "slideInFromBottom 300ms forwards",
+      slideOutToBottom: "slideOutToBottom 300ms forwards",
+      slideInFromRight: "slideInFromRight 300ms forwards",
+      slideOutToRight: "slideOutToRight 300ms forwards",
+      slideDownAndFade: "slideDownAndFade 400ms cubic-bezier(0.16, 1, 0.3, 1)",
+      slideLeftAndFade: "slideLeftAndFade 400ms cubic-bezier(0.16, 1, 0.3, 1)",
+      slideUpAndFade: "slideUpAndFade 400ms cubic-bezier(0.16, 1, 0.3, 1)",
+      slideRightAndFade:
+        "slideRightAndFade 400ms cubic-bezier(0.16, 1, 0.3, 1)",
+    },
+    backgroundSize: {
+      "x-small-gradient": "100% 5%",
+      "y-small-gradient": "5% 100%",
+    },
+    bottom: SPACING_DEFAULTS,
+    height: SPACING_DEFAULTS,
+    keyframes: {
+      fadeIn: {
+        "0%": { opacity: "0" },
+        "100%": { opacity: "1" },
+      },
+      fadeOut: {
+        "0%": { opacity: "1" },
+        "100%": { opacity: "0" },
+      },
+      slideInFromBottom: {
+        "0%": { transform: "translateY(100%)" },
+        "100%": { transform: "translateY(0)" },
+      },
+      slideOutToBottom: {
+        "0%": { transform: "translateY(0)" },
+        "100%": { transform: "translateY(100%)" },
+      },
+      slideInFromRight: {
+        "0%": { transform: "translateX(100%)" },
+        "100%": { transform: "translateX(0)" },
+      },
+      slideOutToRight: {
+        "0%": { transform: "translateX(0)" },
+        "100%": { transform: "translateX(100%)" },
+      },
+      slideDownAndFade: {
+        from: { opacity: "0", transform: "translateY(-2px)" },
+        to: { opacity: "1", transform: "translateY(0)" },
+      },
+      slideLeftAndFade: {
+        from: { opacity: "0", transform: "translateX(2px)" },
+        to: { opacity: "1", transform: "translateX(0)" },
+      },
+      slideUpAndFade: {
+        from: { opacity: "0", transform: "translateY(2px)" },
+        to: { opacity: "1", transform: "translateY(0)" },
+      },
+      slideRightAndFade: {
+        from: { opacity: "0", transform: "translateX(-2px)" },
+        to: { opacity: "1", transform: "translateX(0)" },
+      },
+    },
+    left: SPACING_DEFAULTS,
+    maxHeight: {
+      "90vh": "90vh",
+      "80vh": "80vh",
+    },
+    right: SPACING_DEFAULTS,
+    top: SPACING_DEFAULTS,
+    width: SPACING_DEFAULTS,
+  },
+};
